requirements.txt
Django==4.0.6
Pillow==9.2.0
django-admin-sortable2==2.1
django-admin-interface==0.19.1